requirements.txt
gradio>=4.0.0
SpeechRecognition>=3.8.1
python-dotenv>=1.0.0
requests>=2.28.0
